Move head tag utlis from src/client.js to src/core/DOMUtlis.js
